+++ src/main/java/kg/itacademy/service/impl/CourseServiceImpl.java
@@ -2,7 +2,6 @@ package kg.itacademy.service.impl;
 
 import kg.itacademy.converter.CourseConverter;
 import kg.itacademy.entity.Course;
-import kg.itacademy.entity.Like;
 import kg.itacademy.exception.ApiFailException;
 import kg.itacademy.model.CourseModel;
 import kg.itacademy.repository.CourseRepository;
@@ -23,6 +22,8 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     private final UserService userService;
 
+    private final CourseConverter CONVERTER = new CourseConverter();
+
     @Override
     public Course save(Course course) {
         validateEmail(course);
@@ -34,8 +35,8 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     @Override
     public CourseModel createCourse(CourseModel courseModel) {
-        save(new CourseConverter().convertFromModel(courseModel));
-        return courseModel;
+        Course course = save(CONVERTER.convertFromModel(courseModel));
+        return CONVERTER.convertFromEntity(course);
     }
 
     @Override
@@ -45,7 +46,7 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     @Override
     public CourseModel getCourseModelById(Long id) {
-        return new CourseConverter().convertFromEntity(getById(id));
+        return CONVERTER.convertFromEntity(getById(id));
     }
 
     @Override
@@ -55,9 +56,8 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     @Override
     public List<CourseModel> getAllCourseModel() {
-        CourseConverter converter = new CourseConverter();
         return getAll().stream()
-                .map(converter::convertFromEntity)
+                .map(CONVERTER::convertFromEntity)
                 .collect(Collectors.toList());
     }
 
@@ -72,64 +72,47 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     @Override
     public CourseModel updateCourse(CourseModel courseModel) {
-        update(new CourseConverter().convertFromModel(courseModel));
+        update(CONVERTER.convertFromModel(courseModel));
         return courseModel;
     }
 
     @Override
     public List<CourseModel> getAllByCourseName(String courseName) {
-        List<Course> courses = courseRepository.findAllByCourseName(courseName);
-        if (!courses.isEmpty()) {
-            CourseConverter converter = new CourseConverter();
-            return courses.stream()
-                    .map(converter::convertFromEntity)
-                    .collect(Collectors.toList());
-        }
-        return null;
+        return courseRepository.findAllByCourseName(courseName).stream()
+                .map(CONVERTER::convertFromEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public List<CourseModel> getAllByCourseCategoryName(String categoryName) {
-        List<Course> courses = courseRepository.findAllByCategoryName(categoryName);
-        if (!courses.isEmpty()) {
-            CourseConverter converter = new CourseConverter();
-            return courses.stream()
-                    .map(converter::convertFromEntity)
-                    .collect(Collectors.toList());
-        }
-        return null;
+        return courseRepository.findAllByCategoryName(categoryName).stream()
+                .map(CONVERTER::convertFromEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public List<CourseModel> getAllByCategoryId(Long id) {
-        List<Course> courses = courseRepository.findAllByCategory_Id(id);
-        if (!courses.isEmpty()) {
-            CourseConverter converter = new CourseConverter();
-            return courses.stream()
-                    .map(converter::convertFromEntity)
-                    .collect(Collectors.toList());
-        }
-        return null;
+        return courseRepository.findAllByCategory_Id(id).stream()
+                .map(CONVERTER::convertFromEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public List<CourseModel> getAllByUserId(Long userId) {
-        List<Course> courses = courseRepository.findAllByUser_Id(userId);
-        if (!courses.isEmpty()) {
-            CourseConverter converter = new CourseConverter();
-            return courses.stream()
-                    .map(converter::convertFromEntity)
-                    .collect(Collectors.toList());
-        }
-        return null;
+        return courseRepository.findAllByUser_Id(userId).stream()
+                .map(CONVERTER::convertFromEntity)
+                .collect(Collectors.toList());
     }
 
     @Override
     public CourseModel deleteCourseById(Long id) {
         Course deleteCourse = getById(id);
 
+        if (deleteCourse == null)
+            throw new ApiFailException("Course by user id " + id + " not found");
+
         courseRepository.delete(deleteCourse);
-        return new CourseConverter().convertFromEntity(deleteCourse);
+        return CONVERTER.convertFromEntity(deleteCourse);
     }
 
     private void validateEmail(Course course) {
@@ -181,9 +164,6 @@ public class CourseServiceImpl implements CourseService, VariableValidation<Cour
 
     @Override
     public void validateVariablesForNullOrIsEmptyUpdate(Course course) {
-        if (course.getCategory() != null && course.getCategory().getId() == null)
-            throw new ApiFailException("Category is not filled");
-
         if (course.getCourseName() != null && course.getCourseName().isEmpty())
             throw new ApiFailException("Course name is not filled");
